% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_add_empty_line.R
\name{body_add_empty_line}
\alias{body_add_empty_line}
\title{Add Empty Lines to a Word Document}
\usage{
body_add_empty_line(doc, n = 1)
}
\arguments{
\item{doc}{An \code{rdocx} object representing the Word document.}

\item{n}{An integer specifying the number of empty lines to add. Default is 1.}
}
\value{
The modified \code{rdocx} object with the added empty lines.
}
\description{
This function adds a specified number of empty lines to a Word document using the \code{officer} package.
}
\details{
The function iterates through the specified number of empty lines and adds them to the document. Each empty line is added as a new paragraph with an empty string.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(officer)
doc <- read_docx()
doc <- body_add_empty_line(doc, n = 3) # 添加3个空白行
print(doc, target = "empty_lines_example.docx")
\dontshow{\}) # examplesIf}
}
