% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm2inche.R
\name{cm2inche}
\alias{cm2inche}
\title{Convert centimeters to inches}
\usage{
cm2inche(x)
}
\arguments{
\item{x}{A numeric value representing the length in centimeters.}
}
\value{
A numeric value representing the length in inches.
}
\description{
This function converts a numeric value from centimeters to inches.
}
\examples{
cm2inche(10)
cm2inche(25.4)

}
