% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_font_size.R
\name{get_font_size}
\alias{get_font_size}
\title{Convert Font Size to Points}
\usage{
get_font_size(font_size = "小四")
}
\arguments{
\item{font_size}{A numeric value representing the font size in points, or a character string
corresponding to a predefined font size (e.g., "小四", "三号"). Multiple values can be provided
as a vector.}
}
\value{
A numeric value or a vector of numeric values representing the font size(s) in points.
}
\description{
This function converts a given font size, specified either as a numeric value (in points) or
as a descriptive character value (e.g., "小四", "三号"), into its corresponding point size.
The function supports both single and multiple font size conversions by using an internal
mapping table to match descriptive font sizes to numeric point values.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_font_size(12) # returns 12
get_font_size("小四") # returns 12
get_font_size(c(11, "小四", "14", "四号")) # returns corresponding point sizes
\dontshow{\}) # examplesIf}
}
