% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infor_with_2_columns.R
\name{get_infor_with_2_columns}
\alias{get_infor_with_2_columns}
\title{Create a Two-Column Information Table}
\usage{
get_infor_with_2_columns(
  x,
  y,
  y_border_color = "black",
  y_border_style = "solid",
  y_border_width = 1,
  x_align = "center",
  y_align = "center",
  x_fontsize = 14,
  y_fontsize = x_fontsize,
  font_family_ch = "宋体",
  font_family_en = "Times New Roman",
  font_family_cs = font_family_en,
  cell_height = cm2inche(c(1, 1, 1, 1)),
  cell_width = cm2inche(c(2, 5)),
  cell_unit = "in",
  padding_top = 0,
  padding_bottom = 0,
  padding_left = 0,
  padding_right = 0
)
}
\arguments{
\item{x}{contents for 1st column.}

\item{y}{contents for 2nd column.}

\item{y_border_color}{Border color for 2nd column (default is "black").}

\item{y_border_style}{Border style for 2nd column (default is "solid"). see \link[officer]{fp_border}.}

\item{y_border_width}{Border width for 2nd column (default is 1). see \link[officer]{fp_border}.}

\item{x_align}{1st column. text alignment - a single character value, expected value is one of 'left', 'right', 'center', 'justify'. 'distributed’ alignment is not possible.}

\item{y_align}{2nd column. text alignment - a single character value, expected value is one of 'left', 'right', 'center', 'justify'. 'distributed' alignment is not possible.}

\item{x_fontsize}{Font size for 1st column text.}

\item{y_fontsize}{Font size for 2nd column text.}

\item{font_family_ch}{Font family for Chinese characters (default is "宋体").}

\item{font_family_en}{Font family for English characters (default is "Times New Roman").}

\item{font_family_cs}{Font family for complex script Unicode range.}

\item{cell_height}{Cell height in inches (default is cm2inche(c(1,1,1,1))). Temporarily invalid settings.}

\item{cell_width}{Cell width in inches (default is cm2inche(c(2,5))).}

\item{cell_unit}{Unit for cell width, one of "in", "cm", "mm".}

\item{padding_top}{Padding at the top of the cell (default is 0). Adjust the cell height of the table.}

\item{padding_bottom}{Padding at the bottom of the cell (default is 0). Adjust the cell height of the table.}

\item{padding_left}{Padding on the left side of the cell (default is 0).}

\item{padding_right}{Padding on the right side of the cell (default is 0).}
}
\value{
A \code{flextable} object representing the customized two-column table.
}
\description{
This function generates a two-column table using the \code{flextable} package,
where each column contains text with customizable formatting options.
1st column (\code{x}) typically represents descriptions, and 2nd column (\code{y})
represents corresponding values. The table allows customization of font sizes,
alignment, borders, cell dimensions, and padding.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- c("Name", "School", "Grade", "Date")
y <- c("Xiaoming", "Yangguang School", "Grade 3", "2024-08-28")
table <- get_infor_with_2_columns(x, y)
read_docx() \%>\% body_add_flextable(table, align = "center") \%>\%
  print(target = "infor_table_Example1.docx")
\dontshow{\}) # examplesIf}
}
