% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paragraph2.R
\name{get_paragraph2}
\alias{get_paragraph2}
\title{Create a Formatted Paragraph}
\usage{
get_paragraph2(
  ...,
  tab = "\\t",
  n_tab = 1,
  font_size = 12,
  font_family_ch = "宋体",
  font_family_en = "Times New Roman",
  font_family_cs = NULL,
  bold = NULL,
  italic = NULL,
  color = "black",
  vertical_align = "baseline",
  underlined = NULL,
  shading_color = "transparent",
  text_align = "justify",
  padding_left = 0,
  padding_right = 0,
  padding_top = 0,
  padding_bottom = 0,
  line_spacing = 1.25
)
}
\arguments{
\item{...}{Some text representing the content of the paragraph.}

\item{tab}{A character pasted at the beginning of text text. Default is a tab character (\code{"\\t"}).}

\item{n_tab}{An integer specifying the number of tab characters.}

\item{font_size}{A numeric value or vector indicating the font size. Default is 12.}

\item{font_family_ch}{A character string specifying the font family for Chinese characters. Default is "宋体".}

\item{font_family_en}{single character value. Specifies the font to be used to format characters in the Unicode range (U+0000-U+007F). Default is "Times New Roman".}

\item{font_family_cs}{optional font to be used to format characters in a complex script Unicode range. For example, Arabic text might be displayed using the "Arial Unicode MS" font.}

\item{bold}{is bold. NULL or a numeric vector with same length to all text.}

\item{italic}{is italic. NULL or a numeric vector with same length to all text.}

\item{color}{text color. a character or a vector with same length to all text. Default is "black".}

\item{vertical_align}{single character value or a vector specifying font vertical alignments. Expected value is one of the following : default 'baseline' or 'subscript' or 'superscript'.}

\item{underlined}{is underlined. NULL or a numeric vector with same length to all text.}

\item{shading_color}{shading color - a single character value or a vector specifying a valid color (e.g. "#000000" or "black").}

\item{text_align}{text alignment - a single character value, expected value is one of 'left', 'right', 'center', 'justify'.}

\item{padding_left}{A numeric value specifying the left padding of the paragraph.}

\item{padding_right}{A numeric value specifying the right padding of the paragraph.}

\item{padding_top}{A numeric value specifying the top padding of the paragraph.}

\item{padding_bottom}{A numeric value specifying the bottom padding of the paragraph.}

\item{line_spacing}{line spacing, 1 is single line spacing, 2 is double line spacing.}
}
\value{
A \code{fpar} object representing the formatted paragraph.
}
\description{
This function generates a formatted paragraph for use in Word documents. It allows customization of text content, font properties, alignment, padding, and line spacing.
}
\details{
This function uses \code{fp_text} to set text formatting options such as font size, color, and bold style, and \code{fp_par} to set paragraph formatting options including alignment, padding, and line spacing. The result is a \code{fpar} object that can be used with functions from the \code{officer} package to create rich text content in Word documents.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(officer)
text <- "《黑神话：悟空》是一款由中国游戏公司 Game Science 开发的动作角色扮演游戏（ARPG），灵感来自中国古典名著《西游记》。游戏以孙悟空为主角，结合了传统的中国神话元素和现代游戏设计理念，呈现出一个丰富的开放世界。"
para <- get_paragraph2(text, tab = " ", n_tab = 8)
read_docx() \%>\% body_add_fpar(para) \%>\% print("formatted_paragraph.docx")

text1 <- "《"
text2 <- "黑神话：悟空"
text3 <- "》是一款由中国游戏公司"
text4 <- "Game Science"
text5 <- "开发的动作角色扮演游戏（ARPG），灵感来自中国古典名著《西游记》。游戏以孙悟空为主角，结合了传统的中国神话元素和现代游戏设计理念，呈现出一个丰富的开放世界。"
text <- c(text1,text2,text3,text4,text5)

# Except for the parameter `tab` and `n_tab`, all parameters in this example can be single values or a vector.
para <- get_paragraph2(text, tab = "\t", n_tab = 1,
                       font_size = c(12,13,"四号",15,"二号"),
                       font_family_ch = "宋体",
                       font_family_en = "Times New Roman",
                       bold = c(2,4),
                       italic = c(4),
                       color = c("black","blue","red","green","purple"),
                       vertical_align = "baseline",
                       underlined = c(4,5),
                       shading_color = c("transparent","red","transparent","transparent","transparent")
                       )
read_docx() \%>\% body_add_fpar(para) \%>\% print("formatted_paragraph2.docx")
\dontshow{\}) # examplesIf}
}
