% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_page_layout.R
\name{set_A3_cover_layout}
\alias{set_A3_cover_layout}
\title{Set A3 Cover Page Layout}
\usage{
set_A3_cover_layout(
  page_width = 11.7,
  page_height = 16.5,
  top = 0,
  bottom = 0,
  left = 0,
  right = 0,
  header = 0,
  footer = 0
)
}
\arguments{
\item{page_width}{The width of the page in inches. Default is 11.7 inches (A3 width).}

\item{page_height}{The height of the page in inches. Default is 16.5 inches (A3 height).}

\item{top}{The top margin in inches. Default is 0.}

\item{bottom}{The bottom margin in inches. Default is 0.}

\item{left}{The left margin in inches. Default is 0.}

\item{right}{The right margin in inches. Default is 0.}

\item{header}{The header margin in inches. Default is 0.}

\item{footer}{The footer margin in inches. Default is 0.}
}
\value{
A \code{prop_section} object containing the layout properties for the A3 cover page.
}
\description{
This function configures a section layout for an A3-sized cover page in a Word document. It sets the page dimensions, margins, headers, and footers using the standard A3 size.
}
\details{
This function is a wrapper around \code{set_page_layout}, specifically for creating an A3-sized cover page layout. It sets the page size to the A3 standard dimensions and allows for customization of the page margins, headers, and footers.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(officer)
cover_layout <- set_A3_cover_layout()
doc <- read_docx() \%>\%
  body_add_par("A3 Cover Page Title", style = "heading 1") \%>\%
  body_set_default_section(cover_layout) \%>\%
  print(target = "A3_cover_page.docx")
\dontshow{\}) # examplesIf}
}
