% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_page_layout.R
\name{set_A4_cover_layout}
\alias{set_A4_cover_layout}
\title{Set A4 Cover Page Layout}
\usage{
set_A4_cover_layout(
  page_width = 8.27,
  page_height = 11.69,
  top = 0,
  bottom = 0,
  left = 0,
  right = 0,
  header = 0,
  footer = 0
)
}
\arguments{
\item{page_width}{The width of the page in inches. Default is 8.27 inches, which is the standard width for A4 paper.}

\item{page_height}{The height of the page in inches. Default is 11.69 inches, which is the standard height for A4 paper.}

\item{top}{The top margin in inches. Default is 0.}

\item{bottom}{The bottom margin in inches. Default is 0.}

\item{left}{The left margin in inches. Default is 0.}

\item{right}{The right margin in inches. Default is 0.}

\item{header}{The header margin in inches. Default is 0.}

\item{footer}{The footer margin in inches. Default is 0.}
}
\value{
A \code{prop_section} object with the layout properties for the A4 cover page.
}
\description{
This function sets up the layout for an A4-sized cover page in a Word document. It configures the page dimensions and margins specifically for A4 paper size.
}
\details{
This function is a wrapper around the \code{set_page_layout} function, pre-configured for A4-sized paper. It is useful for generating cover pages in A4 format with specific margins and layout settings.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(officer)
cover_layout <- set_A4_cover_layout()
doc <- read_docx() \%>\%
  body_add_par("Cover Page Title", style = "heading 1") \%>\%
  body_set_default_section(cover_layout) \%>\%
  print(target = "A4_cover_page.docx")

# Using the image as the cover
doc <- read_docx() \%>\%
  body_set_default_section( cover_sect_properties ) \%>\%
  body_add_img(src = "./input/cover.jpg", width = 8.27, height = 11.69) \%>\%
  body_end_block_section(value = block_section(property = cover_sect_properties)) # 分节设置
\dontshow{\}) # examplesIf}
}
