% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_page_layout.R
\name{set_B5_cover_layout}
\alias{set_B5_cover_layout}
\title{Set B5 Cover Page Layout}
\usage{
set_B5_cover_layout(
  page_width = 6.93,
  page_height = 9.84,
  top = 0,
  bottom = 0,
  left = 0,
  right = 0,
  header = 0,
  footer = 0
)
}
\arguments{
\item{page_width}{The width of the page in inches. Default is 6.93 inches (B5 width).}

\item{page_height}{The height of the page in inches. Default is 9.84 inches (B5 height).}

\item{top}{The top margin in inches. Default is 0.}

\item{bottom}{The bottom margin in inches. Default is 0.}

\item{left}{The left margin in inches. Default is 0.}

\item{right}{The right margin in inches. Default is 0.}

\item{header}{The header margin in inches. Default is 0.}

\item{footer}{The footer margin in inches. Default is 0.}
}
\value{
A \code{prop_section} object containing the layout properties for the B5 cover page.
}
\description{
This function configures a section layout for a B5-sized cover page in a Word document. It sets the page dimensions, margins, headers, and footers using the standard B5 size.
}
\details{
This function is a wrapper around \code{set_page_layout}, specifically for creating a B5-sized cover page layout. It sets the page size to the B5 standard dimensions and allows for customization of the page margins, headers, and footers.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(officer)
cover_layout <- set_B5_cover_layout()
doc <- read_docx() \%>\%
  body_add_par("B5 Cover Page Title", style = "heading 1") \%>\%
  body_set_default_section(cover_layout) \%>\%
  print(target = "B5_cover_page.docx")
\dontshow{\}) # examplesIf}
}
